+++ src/tools/voxedit/modules/voxedit-util/SceneManager.h
@@ -570,6 +570,7 @@ public:
 	bool nodeSetColor(int nodeId, uint8_t palIdx, const color::RGBA &color);
 	void nodeResize(int nodeId, const voxel::Region &region);
 	void nodeResize(int nodeId, const glm::ivec3 &size);
+	void nodeBakeTransform(int nodeId);
 	/**
 	 * @brief If a type of a palette color changes its transparency state, we have to update the voxels
 	 * in the volume that are using this color. This is because we separate the color and the alpha voxels